+++ maze/__init__.py
@@ -1,5 +1,5 @@
-__all__ = ['adsorbate', 'cif_download', 'extra_framework_maker', 'index_mapper', 'io_zeotype', 'perfect_zeolite',
+__all__ = ['adsorbate', 'cif_download', 'extra_framework_maker', 'index_mapper', 'io_zeolite.py', 'perfect_zeolite',
            'silanol', 'zeolite', 'zeolite_db']
-from maze import adsorbate, cif_download, extra_framework_maker, index_mapper, io_zeotype, perfect_zeolite, silanol
+from maze import adsorbate, cif_download, extra_framework_maker, index_mapper, io_zeolite, perfect_zeolite, silanol
 from maze import zeolite, zeolite_db
 from maze.zeolite import Zeolite
